Get weather data for use in epicrop modelling

This function is a wrapper for the nasapower::get_power() or chirps::get_chirps()/chirps::get_chirts() functions with predefined parameters suitable for use in epicrop.

get_wth(lonlat, dates, duration, source = "nasapower")

Arguments

lonlat
A numeric vector of geographic coordinates for a cell or region entered as x, y coordinates.

dates
A character vector of start and end dates in that order.

duration
A numeric value indicating how many days a single growing season being simulated is to be. If this is used, only the start date will be used and the end date will be ignored if supplied. This must match the duration parameter value passed along to SEIR() or any of the predict family of functions.

source
A character string value that denotes whether to use the NASA POWER data (nasapower), the default, or CHIRPS/CHIRTS data, chirps, for rainfall, temperature and relative humidity from the Climate Hazards Data center at UC Santa Barbara, https://www.chc.ucsb.edu/data. See parameter details for more discussion.

Value

A data.table::data.table() of weather data, dates and geolocation information (LAT/LON values) suitable for use in epicrop with the following columns:

Field Name	Value

YYYYMMDD	Date as Year Month Day (ISO8601)

DOY	Consecutive day of year, commonly called "Julian date"

TEMP	Mean daily temperature (°C)

RHUM	Mean daily relative humidity (%)

RAIN	Mean daily rainfall (mm)

LAT	Latitude of area of interest

LON	Longitude of area of interest

source - Select your weather data source of choice

Access to three databases is provided by this function through the use of two possible parameter values, nasapower or chirps.

nasapower: the default database is the NASA POWER database, https://power.larc.nasa.gov, which can supply all parameters required for epicrop to run. The global data are available from Jan. 1, 1983 to near-present at a 0.5 x 0.5 arc degree resolution for meteorology data. Using this will request and return all values necessary to use epicrop.

chirps: CHIRPS is quasi-global (50°S – 50°N) high-resolution (0.05 arc-degrees) rainfall data set, which incorporates satellite imagery and in-situ station data to create gridded rainfall time series for trend analysis and seasonal drought monitoring. CHIRTS is a quasi-global (60°S – 70°N), high-resolution data set of daily maximum and minimum temperatures. For more details on CHIRPS and CHIRTS data please visit the official homepage https://chc.ucsb.edu/data.

By using chirps, the function will return the precipitation, temperature and relative humidity from the CHIRPS and CHIRTS data in a single data.table for you.

The POWER database covers a longer period of time but the spatial resolution is not as small. However, it is faster to respond than the CHIRPS and CHIRTS data. Differences in the data are likely to be minimal and have little effect on the epicrop models.

Author

Adam H. Sparks

Examples

if (FALSE) { # interactive()

# get weather for IRRI Zeigler Experiment Station in wet season 2000 from the
# default NASA POWER data.
power <- get_wth(
  lonlat = c(121.25562, 14.6774),
  dates = c("2000-06-30", "2000-12-31")
)

# get 120 days of weather for IRRI Zeigler Experiment Station in wet season
# 2000 by specifying the duration but not the end-date and specifying to use
# POWER data.
power <- get_wth(
  lonlat = c(121.25562, 14.6774),
  dates = "2000-06-30",
  duration = 120,
  source = "nasapower"
)

# get the same data from CHIRPS and CHIRTS data
chirps <- get_wth(
  lonlat = c(121.25562, 14.6774),
  dates = "2000-06-30",
  duration = 120,
  source = "chirps"
)
}