Predict rice tungro disease incidence

A dynamic mechanistic simulation of tungro disease of rice, causal agents Rice Tungro Spherical Virus and Rice Tungro Bacilliform Virus. The model is driven by daily weather data, which can easily be accessed using get_wth() to download weather data from NASA POWER using nasapower.

predict_tungro(wth, emergence)

Arguments

wth

Weather data with a daily time-step, normally NASA POWER data from get_wth(), but anybase::data.frame() object that has the following properly named columns in them will work.

Field Name	Value

YYYYMMDD	Date as Year Month Day (ISO8601)

DOY	Consecutive day of year, commonly called "Julian date"

TEMP	Mean daily temperature (°C)

RHUM	Mean daily relative humidity (%)

RAIN	Mean daily rainfall (mm)

LAT	Optional latitude of weather observation. See LAT/LON Note.

LON	Optional longitude of weather observation. See LAT/LON Note.

emergence

Expected date of crop emergence

Value

A data.table::data.table() of disease intensity and infection sites. See SEIR() for a full description of the column values.

Details

The model represents site size as 1 rice plant.

Default values for this disease model are derived from Table 2 (Savary et al. 2012).

Note

Adapted from cropsim package version 0.2.0-5 by Adam H. Sparks, Department of Primary Industries and Regional Development, WA, AU. Original model development: Serge Savary & Rene Pangga (IRRI). Original R implementation by Robert J. Hijmans, Rene Pangga, & Jorrel Aunario (IRRI).

If the wth object provides LAT and LON columns, these will be included in the output for mapping purposes. Both values must be present. These columns are provided by default when using get_wth().

References

Ling, K.C., and Tiongco, E.R., 1976. Effect of temperature on the transmission of rice tungro virus by Nephotettix virescens. Philippine Phytopathology 11:46-57.

Ling, K.C., Palomar, M.K., 1966. Studies on rice plants infected with the tungro virus at different ages. Philippines Agriculturist 50:165-177.

Rivera, C.T. and Ou, S.H., 1965. Leafhopper transmission of tungro disease of rice. Plant Disease Reporter 49:127-131.

Savary, S., Nelson, A., Willocquet, L., Pangga, I., and Aunario, J. Modeling and mapping potential epidemics of rice diseases globally. Crop Protection, Volume 34, 2012, Pages 6-17, ISSN 0261-2194 DOI: doi:10.1016/j.cropro.2011.11.009 .

Tiongco, E.R., Cabunagan, R.C., Flores, Z.M., Hibino, H., and Koganezawa, H., 1993. Serological monitoring of rice tungro disease development in the field: its implication in disease management. Plant Disease 77:877-882. DOI: doi:10.1094/PD-77-0877 .

See also

Author

Serge Savary, Ireneo Pangga, Robert Hijmans, Jorrel Khalil Aunario and Adam H. Sparks

Examples

if (FALSE) { # interactive()

# get weather for IRRI Zeigler Experiment Station in wet season 2000
wth <- get_wth(
  lonlat = c(121.25562, 14.6774),
  dates = c("2000-06-30", "2000-12-31")
)
tg <- predict_tungro(wth, emergence = "2000-07-01")
plot(x = tg$dates, y = tg$intensity, type = "l")
}