Susceptible-Exposed-Infectious-Removed (SEIR) model framework

This function is originally used by specific disease models in EPIRICE to model disease severity of several rice diseases. Given proper values it can be used with other pathosystems as well.

SEIR(
  wth,
  emergence,
  onset,
  duration,
  rhlim = 90,
  rainlim = 5,
  H0,
  I0 = 1,
  RcA,
  RcT,
  RcW,
  RcOpt,
  p,
  i,
  Sx,
  a,
  RRS,
  RRG
)

Arguments

wth
a data frame of weather on a daily time-step containing data with the following field names.

YYYYMMDD:	Date as Year Month Day (ISO8601).

DOY:	Consecutive day of year, commonly called "Julian date".

TEMP:	Mean daily temperature (°C).

RHUM:	Mean daily temperature (°C).

RAIN:	Mean daily rainfall (mm).

emergence
expected date of plant emergence entered in YYYY-MM-DD format. From Table 1 Savary et al. 2012.

onset
expected number of days until the onset of disease after emergence date. From Table 1 Savary et al. 2012.

duration
simulation duration (growing season length). From Table 1 Savary et al. 2012.

rhlim
relative humidity value threshold to decide whether leaves are wet or not. Savary et al. 2012 used 90%. From Table 1 Savary et al. 2012.

rainlim
rainfall amount (mm) threshold to decide whether leaves are wet or not. Savary et al. 2012 used 5mm. From Table 1 Savary et al. 2012.

H0
initial number of plant's healthy sites. From Table 1 Savary et al. 2012.

I0
initial number of infective sites. From Table 1 Savary et al. 2012.

RcA
modifier for Rc for crop age. From Table 1 Savary et al. 2012.

RcT
modifier for Rc for temperature. From Table 1 Savary et al. 2012.

RcW
modifier for Rc for wetness. From Table 1 Savary et al. 2012.

RcOpt
potential basic infection rate corrected for removals. From Table 1 Savary et al. 2012.

p
duration of latent period. From Table 1 Savary et al. 2012.

i
duration of infectious period. From Table 1 Savary et al. 2012.

Sx
maximum number of sites. From Table 1 Savary et al. 2012.

a
aggregation coefficient. From Table 1 Savary et al. 2012.

RRS
relative rate of physiological senescence. From Table 1 Savary et al. 2012.

RRG
relative rate of growth. From Table 1 Savary et al. 2012.

Value

A data.table containing the following columns

simday:	Zero indexed day of simulation that was run.

dates:	Date of simulation.

sites:	Total number of sites present on day "x".

latent:	Number of latent sites present on day "x".

infectious:	Number of infectious sites present on day "x".

removed:	Number of removed sites present on day "x".

senesced:	Number of senesced sites present on day "x".

rateinf:	Rate of infection.

rtransfer:	Rate of transfer from latent to infectious sites.

rgrowth:	Rate of growth of healthy sites.

rsenesced:	Rate of senescence of healthy sites.

rlex:	Rate of lesion expansion.

diseased:	Number of diseased (latent + infectious + removed) sites.

severity:	Disease severity or incidence (for tungro).

lat:	Latitude value as provided by wth object.

lon:	Longitude value as provided by wth object.

Details

SEIR is called by the following specific disease modelling functions:

predict_bacterial_blight,

predict_brown_spot,

predict_leaf_blast,

predict_sheath_blight,

predict_tungro

References

Savary, S., Nelson, A., Willocquet, L., Pangga, I., and Aunario, J. Modeling and mapping potential epidemics of rice diseases globally. Crop Protection, Volume 34, 2012, Pages 6-17, ISSN 0261-2194 DOI: http://dx.doi.org/10.1016/j.cropro.2011.11.009.

Savary, S., Stetkiewicz, S., Brun, F., and Willocquet, L. Modelling and Mapping Potential Epidemics of Wheat Diseases—Examples on Leaf Rust and Septoria Tritici Blotch Using EPIWHEAT. European Journal of Plant Pathology 142, no. 4 (August 1, 2015): 771–90. DOI: https://doi.org/10.1007/s10658-015-0650-7.

Author

Serge Savary, Ireneo Pangga, Robert Hijmans, Jorrel Khalil Aunario, Adam H. Sparks, Aji Sukarta

Examples

# \donttest{
# get weather for IRRI Zeigler Experiment Station in wet season 2000
wth <- get_wth(
  lonlat = c(121.25562, 14.6774),
  dates = c("2000-06-30", "2000-12-31")
)

# provide suitable values for brown spot severity
RcA <-
  cbind(0:6 * 20, c(0.35, 0.35, 0.35, 0.47, 0.59, 0.71, 1.0))
RcT <-
  cbind(15 + (0:5) * 5, c(0, 0.06, 1.0, 0.85, 0.16, 0))
RcW <- cbind(0:8 * 3,
             c(0, 0.12, 0.20, 0.38, 0.46, 0.60, 0.73, 0.87, 1.0))
emergence <- "2000-07-15"

(SEIR(
  wth = wth,
  emergence = emergence,
  onset = 20,
  duration = 120,
  RcA = RcA,
  RcT = RcT,
  RcW = RcW,
  RcOpt = 0.61,
  p =  6,
  i = 19,
  H0 = 600,
  a = 1,
  Sx = 100000,
  RRS = 0.01,
  RRG = 0.1
))

#>      simday      dates      sites   latent infectious  removed    senesced
#>   1:      0 2000-07-14   600.0000    0.000      0.000   0.0000     0.00000
#>   2:      1 2000-07-15   653.6400    0.000      0.000   0.0000     6.00000
#>   3:      2 2000-07-16   712.0404    0.000      0.000   0.0000    12.53640
#>   4:      3 2000-07-17   775.6170    0.000      0.000   0.0000    19.65680
#>   5:      4 2000-07-18   844.8209    0.000      0.000   0.0000    27.41297
#>  ---                                                                      
#> 117:    116 2000-11-07 85727.9827 1597.542   1993.769 191.2783 52375.66862
#> 118:    117 2000-11-08 85734.5312 1597.542   1958.359 226.6876 53268.35769
#> 119:    118 2000-11-09 84730.7315 2228.588   2289.768 268.0599 54167.07534
#> 120:    119 2000-11-10 84721.7757 1688.911   2779.577 317.9281 55064.25088
#> 121:    120 2000-11-11 84705.1565 1003.827   3406.375 376.2142 55969.75472
#>       rateinf rtransfer   rgrowth  rsenesced diseased severity      lat
#>   1:    0.000    0.0000  59.64000   6.000000    0.000 0.000000 14.67741
#>   2:    0.000    0.0000  64.93675   6.536400    0.000 0.000000 14.67741
#>   3:    0.000    0.0000  70.69703   7.120404    0.000 0.000000 14.67741
#>   4:    0.000    0.0000  76.96012   7.756170    0.000 0.000000 14.67741
#>   5:    0.000    0.0000  83.76837   8.448209    0.000 0.000000 14.67741
#>  ---                                                                   
#> 117:    0.000    0.0000 899.23753 892.689065 3782.589 4.020756 14.67741
#> 118: 1003.827  372.7810 898.74479 898.717650 3782.589 3.982399 14.67741
#> 119:    0.000  539.6772 888.21978 897.175543 4786.416 5.062635 14.67741
#> 120:    0.000  685.0837 888.88464 905.503838 4786.416 5.010062 14.67741
#> 121:    0.000    0.0000 890.11801 912.577369 4786.416 4.948868 14.67741
#>           lon
#>   1: 121.2556
#>   2: 121.2556
#>   3: 121.2556
#>   4: 121.2556
#>   5: 121.2556
#>  ---         
#> 117: 121.2556
#> 118: 121.2556
#> 119: 121.2556
#> 120: 121.2556
#> 121: 121.2556

# }