Susceptible-Exposed-Infectious-Removed (SEIR) model framework

This function is originally used by specific disease models in ‘EPIRICE’ to model disease intensity of several rice diseases. Given proper values it can be used with other pathosystems as well.

SEIR(
  wth,
  emergence,
  onset,
  duration,
  rhlim,
  rainlim,
  H0,
  I0,
  RcA,
  RcT,
  RcOpt,
  p,
  i,
  Sx,
  a,
  RRS,
  RRG
)

Arguments

wth
a data.frame of weather on a daily time-step containing data with the following field names.

Field Name	Value

YYYYMMDD	Date as Year Month Day (ISO8601)

DOY	Consecutive day of year, commonly called "Julian date"

TEMP	Mean daily temperature (°C)

RHUM	Mean daily temperature (°C)

RAIN	Mean daily rainfall (mm)

LAT	Optional latitude of weather observation. See LAT/LON Note.

LON	Optional longitude of weather observation. See LAT/LON Note.

emergence
expected date of plant emergence (or transplanting for rice) entered in YYYY-MM-DD format (character). Described in Table 1 of Savary et al. 2012.

onset
expected number of days until the onset of disease after emergence date (day, integer). Described in Table 1 of Savary et al. 2012.

duration
simulation duration i.e., growing season length (day, integer). Described in Table 1 of Savary et al. 2012.

rhlim
relative humidity value threshold to decide whether leaves are wet or not (numeric). Savary et al. 2012 used 90%.

rainlim
rainfall amount (mm) threshold to decide whether leaves are wet or not (numeric). Savary et al. 2012 used 5mm.

H0
initial number of plant's healthy sites (integer). Described in Table 1 of Savary et al. 2012.

I0
initial number of infective sites (integer). Described in Table 1 of Savary et al. 2012.

RcA
modifier for Rc (the basic infection rate corrected for removals) for crop age (numeric vector). Described in Table 1 of Savary et al. 2012.

RcT
modifier for Rc (the basic infection rate corrected for removals) for temperature (numeric vector). Described in Table 1 of Savary et al. 2012.

RcOpt
potential basic infection rate corrected for removals (numeric). Derived from Table 1 of Savary et al. 2012.

p
duration of latent period (day, integer). Described in Table 1 of Savary et al. 2012.

i
duration of infectious period (day, integer). Described in Table 1 of Savary et al. 2012.

Sx
maximum number of sites (integer). Described in Table 1 of Savary et al. 2012.

a
aggregation coefficient, values are from 1 to >1 (numeric). Described in Table 1 of Savary et al. 2012. See further details in a - Aggregation section.

RRS
relative rate of physiological senescence (numeric). Described in Table 1 of Savary et al. 2012.

RRG
relative rate of growth (numeric). Described in Table 1 of Savary et al. 2012.

Value

A data.table::data.table() containing the following columns:

simday

Zero indexed day of simulation that was run

dates

Date of simulation

sites

Total number of sites present on day "x"

latent

Number of latent sites present on day "x"

infectious

Number of infectious sites present on day "x"

removed

Number of removed sites present on day "x"

senesced

Number of senesced sites present on day "x"

ratinf

Rate of infection

rtransfer

Rate of transfer from latent to infectious sites

rgrowth

Rate of growth of healthy sites

rsenesced

Rate of senescence of healthy sites

rlex

Rate of lesion expansion

diseased

Number of diseased (latent + infectious + removed) sites on day "x"

intensity

Proportion of diseased (latent + infectious + removed) sites per total sites not including removed sites on day "x"

lat

Latitude value if provided by wth object

lon

Longitude value if provided by wth object

Note

If the wth object provides LAT and LON columns, these will be included in the output for mapping purposes. Both values must be present. These columns are provided by default when using get_wth().

a - Aggregation

When a is set to 1 the assumption is that that there is no disease aggregation with new infections occurring at random among the healthy sites. When a is greater than 1 there is aggregation in the disease occurrence, the pathogen is unable to access the entire population of healthy sites, which results in disease aggregation. Refer to Savary et al. (2012) for greater detail.

References

Savary, S., Nelson, A., Willocquet, L., Pangga, I., and Aunario, J. Modeling and mapping potential epidemics of rice diseases globally. Crop Protection, Volume 34, 2012, Pages 6-17, ISSN 0261-2194 DOI: http://dx.doi.org/10.1016/j.cropro.2011.11.009.

See also

Author

Adam H. Sparks

Examples

if (FALSE) { # interactive()
# get weather for IRRI Zeigler Experiment Station in wet season 2000
wth <- get_wth(
  lonlat = c(121.25562, 14.6774),
  dates = c("2000-06-30", "2000-12-31")
)

# provide suitable values for brown spot intensity
RcA <-
  cbind(c(0L, 20L, 40L, 60L, 80L, 100L, 120L),
        c(0.35, 0.35, 0.35, 0.47, 0.59, 0.71, 1.0))
RcT <-
  cbind(c(15L, 20L, 25L, 30L, 35L, 40L),
        c(0, 0.06, 1.0, 0.85, 0.16, 0))
emergence <- "2000-07-15"

(SEIR(
  wth = wth,
  emergence = emergence,
  onset = 20,
  duration = 120,
  rhlim = 90,
  rainlim = 5,
  RcA = RcA,
  RcT = RcT,
  RcOpt = 0.61,
  p =  6,
  i = 19,
  H0 = 600,
  I0 = 1,
  a = 1,
  Sx = 100000,
  RRS = 0.01,
  RRG = 0.1
))
}